requirements.txt
## tornado web framework
tornado==5.1.1

## aiomysql driver
aiomysql==0.0.9
PyMySQL==0.8.0

## aioredis
aioredis
msgpack-python==0.5.5

## validator
jsonschema==2.6.0